solucionando errores... y más errores...

+++ app/Http/Controllers/FigurasController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Figura;
 use App\Pintura;
+use App\Procedimiento;
+use DB;
 use Input;
 
 
@@ -17,6 +19,8 @@ class FigurasController extends Controller
      */
     public function index()
     {
+        $figura = Figura::all();
+        return view('post.figuras')->with('figura',$figura);
     }
     public function nuevaminiatura(){
         $figuras = Figura::all();
@@ -74,11 +78,48 @@ class FigurasController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Request $request)
     {
-        //
+        $id_figuras=$request->id_figuras;
+        $figurin=Figura::find($id_figuras);
+           $salida = array(
+                'Nombre' => $figurin->nombre_figuras,
+                'Alianza' => $figurin->alianza_figuras,
+                'Ejercito' => $figurin->ejercito_figuras,
+                'Precio' => $figurin->precio_figuras,
+                'id' => $figurin->$id_figuras,
+            );
+            echo json_encode($salida);
     }
 
+    public function UploadPDF (Request $request, $id_figuras){
+        if ($request->hasFile('normativaspdf')) {
+            $figuras = Figura::find($request->id);
+            $pdf = $request->file('normativaspdf');
+            $ci_extension = $pdf->getClientOriginalExtension();
+            $nombree = time().'pdf.'.$ci_extension;
+            $pdf->storeAs('/normas', $nombree);
+            $figuras->normas = $nombree;
+            $figuras->save();
+            return redirect('/add/figuras');
+
+        }
+    }
+    public function Procedimiento (Request $request, $id_figuras){
+        $procedimento = new Procedimiento();
+        $procedimento->contenido_noticia = $request->input('procedmiento');
+        $procedimento->id_figuras = $id_figuras;
+        $procedimento->save();
+        return redirect('add/figuras');
+    }
+
+    public function mostrarcompleta($id_figuras){
+      $figuras = Figura::where('id_figuras',$id_figuras)->get();
+       $procedimento = DB::table('procedimientos')
+       ->select('procedimientos.contenido_noticia as contenido_noticia, procedimientos.id_figuras as id_figuras')
+       ->join('figuras','figuras.id_figuras','procedimientos.id_figuras')->get();
+       return view('post.figuracompleta', ['figuras'=> $figuras,'procedimento' => $procedimento]);
+    }
     /**
      * Update the specified resource in storage.
      *
@@ -86,9 +127,9 @@ class FigurasController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id_figuras)
+    public function update(Request $request)
     {
-        $figura = Figura::find($id_figuras);
+        $figura = Figura::find($request->input('id'));
         $this->validate($request,  [
             'minifoto' => 'image|mimes:jpeg,png,jpg,gif|max:2048'
         ]);
@@ -104,28 +145,21 @@ class FigurasController extends Controller
        return redirect ('/add/figuras')->with('status', 'La miniatura ha sido añadida.');          
 
     }
-    public function addpinturaf(Request $request, $id_figuras){
-        $pintura_figura = $request->miembro_investidura['investidura_id'];
-        $cont= 0;
-        while($cont < count($investidura_id)){
-            $miembro_investidura = new Miembro_investidura();
-            $miembro_investidura->miembro_id= $miembro->id;
-            $miembro_investidura->investidura_id= $investidura_id[$cont];
-            $miembro_investidura->save();
-            
-            $cont=$cont + 1;
-        }
-    }
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id_figuras)
-    {
-        $figura = Figura::find($id_figuras);
-        $figura->delete();
-        return redirect('/add/figuras') ;
+    public function destroy(Request $request)
+    {   
+        if($request->ajax()){
+            $id_figuras = $request->id;
+            $figura = Figura::find($id_figuras);
+                $figura->delete();
+                return response()->with('message', '¡Figura borrada!');
+                return redirect('/add/figuras');
+            }
+            
+        }
     }
-}
